+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # (MODX)EvolutionCMS.libraries.ddTools changelog
 
 
+## Version 0.37 (2020-05-24)
+* \* `\ddTools::sort2dArray`:
+	* \+ Associative arrays are also supported.
+	* \* Small refactoring.
+
+
 ## Version 0.36 (2020-05-18)
 * \+ `\ddTools::prepareDocData`, `\ddTools::createDocument`, `\ddTools::updateDocument`: Also can take document fields as stdClass, not only as an associative array.
 
@@ -345,4 +351,5 @@
 * \+ The first release.
 
 
+<link rel="stylesheet" type="text/css" href="https://DivanDesign.ru/assets/files/ddMarkdown.css" />
 <style>ul{list-style:none;}</style>